Add files via upload
This is the JavaScript programs, in this project I build a simple calculator for the price of tickets for a theme park. For example, I could use this program to get a ballpark figure for admission costs to a theme park like Disneyland or Universal Studios. My focus in this project is to follow JavaScript topics: accessing the DOM, creating variables, and making basic arithmetic calculations.